requirements.txt
Flask==2.3.3
pandas==2.1.1
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
fpdf2==2.7.6
Werkzeug==2.3.7
Jinja2==3.1.2
